requirements.txt
litellm>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.3.0
chromadb>=0.4.0
elevenlabs>=1.0.0
